Add linkdir, unlinkdir commands
`Entry` now supports an optional property `linkDir`.  It has no semantics for
files, but for directories whose entries have `!! linkDir`, a `linkdir` command
will be returned for creation (instead of `mkdir`) and similarly an `unlinkdir`
command will be returned for deletion.

This feature is intended to support an optimization in which directories that
can be symlinked are not traversed.  The `linkdir` operation distinguishes this
to callers from a simple `mkdir` for an empty directory.

It is an error for callers to provide a directory with `!! entry.linkDir` and
also provide entries that are children of that directory

This feature is only supported for `FSTree.fromEntries` and not
`FSTree.fromFiles