[Cleanup] removals in reverse order then additions
Originally our algorithm was simple. Removal in reverse order, then additions. Unfortunately, it was implemented incorrectly. When we fixed a bug here, we add unneeded complexity… introducing another bug. 2cffc919 fixes the bug we introduced, and obvious that our original algorithm (just implemented correctly) was actually fine if we just implemented it correctly.