fix: Don't mix generated declaration files with ts source files.
TypeScript seems to get confused when a .d.ts and a .ts file of the same
basename are in the same directory. This causes a consuming project to
think it needs to compile the *.ts files that are imported from the
index.d.ts file.

Also did some other maintainance while I was here:
* use yarn instead of npm for script execution.
* enable sourcemaps so consuming projects will see the right source when
  navigating code and so debugging sessions can navigate the typescript code.